+++ camel/utils/__init__.py
@@ -11,10 +11,12 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 # =========== Copyright 2023 @ CAMEL-AI.org. All Rights Reserved. ===========
+
 from .commons import (
     PYDANTIC_V2,
     api_key_required,
     check_server_running,
+    continue_search,
     download_tasks,
     get_first_int,
     get_prompt_template_key_words,
@@ -22,6 +24,8 @@ from .commons import (
     get_task_list,
     print_text_animated,
     role_playing_with_function,
+    summarize_text,
+    text_extract_from_web,
     to_pascal,
 )
 from .token_counting import (
@@ -49,4 +53,7 @@ __all__ = [
     'OpenAITokenCounter',
     'OpenSourceTokenCounter',
     'role_playing_with_function',
+    'continue_search',
+    'summarize_text',
+    'text_extract_from_web',
 ]
